Flume 1.6.0 supporting ElasticSearch 2.2.0
This version of the Flume ElasticSearch sink works with ElasticSearch 2.2.0
but to do so, some libreries used by Flume had to be updated to the newest
versions, for example: elasticsearch (it's obvious) or guava (from 11.0.2
to 19.0). For this reason, many tests do not work retuning a pretty "jar
hell" exception.